Make the text engine's TimerManager class work with SAPI 5 too
- Move TextTimerManager class into engines.base.timer as
  ThreadedTimerManager and adjust backend_text.
- Stop requiring engine.connect() before using timers with the
  manager class. This is an unnecessary limitation.
- Timeout thread.join() after 5 seconds to prevent the main thread
  hanging if timers are misused.
- Use ThreadedTimerManager in SAPI 5 engine backend.
